Move proto conversion to ML
